move "onJobPsfsChanged" callback to "JobPsfsMgr"
